menu_masterserver.cpp:add function to play different sound
This adds a function to play a different sound when a server becomes
available.

The sound hasn't changed, but now the separate function has been added,
the sound can be changed very easily.

The ticket to create the new sound is at
https://github.com/ZetaGlest/zetaglest-data/issues/44

Can be customized by adding the variable to
$HOME/.zetaglest/glestuser.ini

closes #63