allow hosts of custom games to use slots > hardMaxPlayers
This allows the host of a custom game to switch himself to a slot >
mapInfo.hardMaxPlayers for the purpose of being an observer. The 10th
slot is enabled and editable (listBoxControls).

The 8 lines removed from menu_state_custom_game.cpp was some code I
tried last month and forgot to remove.