Travis:add tests for more gcc versions (#199)
* travis:add tests for other versions of gcc

* cmakelists.txt:add c++11 flag by default, doesn't build using gcc-5

the llvm toolchain is needed for the clang-6.0 test.

Recently something changed on travis where adding the llvm toolchain by
using the "addons" directive in .travis.yml fails because
"allow-unauthenticated" isn't
passed, which seems to be the only way to add the repo. Moving that to
the script, and cleaning up the script a little.